+++ packages/melonjs/src/state/stage.js
@@ -1,4 +1,4 @@
-import Camera2d from "./../camera/camera2d.js";
+import Camera2d from "./../camera/camera2d.ts";
 import { Color } from "./../math/color.ts";
 import { eventEmitter, STAGE_RESET } from "../system/event.ts";
 import { renderer } from "./../video/video.js";